Optimize DB performance by adding the expression In(name, values)
Without an In expression in order to resolve a query we had to perform a
recursive `Query.solve_using` call per solution pre predicate. Making
the amount of queries to the database exponential.

By adding the In expression when the solutions to a predicate variable
returns a set of associated values. We can simplify further predicates
by converting the same variable into an In expression.

This makes the amount of queires to the database have a linear relation
with the amount of predicates in a Query object.

Notice in this commit how the values of number of queries in the file
`tests_using_django.py` went down.